+++ src/include/Core/ResUtil.hpp
@@ -9,4 +9,4 @@
 std::string ReadFile(std::string location);
 void LoadModel(Mesh *mesh, std::string location);
 void LoadTexture();
-int GetRamUsage();
+float GetRamUsage();
